Update hash type to uint32_t in object and table files

+++ csrc/table.c
@@ -20,7 +20,7 @@ void freeTable(Table* table) {
 
 static Entry* findEntry(Entry* entries, int capacity,
                         ObjString* key) {
-    uint64_t index = key->hash & (capacity - 1);
+    uint32_t index = key->hash & (capacity - 1);
     Entry* tombstone = NULL;
 
     for (;;) {
@@ -110,7 +110,7 @@ ObjString* tableFindString(Table* table, const char* chars,
                            int length, uint64_t hash) {
     if (table->count == 0) return NULL;
 
-    uint64_t index = hash & (table->capacity - 1);
+    uint32_t index = hash & (table->capacity - 1);
     for (;;) {
         Entry* entry = &table->entries[index];
         if (entry->key == NULL) {
